requirements.txt
numpy<2.2,>=1.24
mss>=9.0.1
Pillow>=10.0.0
pyautogui>=0.9.54
opencv-python>=4.8.0
sounddevice>=0.4.6
scipy>=1.11.0
librosa==0.9.2
setuptools>=65.5.1
keyboard>=0.13.5